favoriteRecipes
Переніс логіку з контроллера та сервіса favoriteRecipes у контролер та сервіс recipes. Видалив файли favoriteRecipesControllers та favoriteRecipesServices
