+++ routes/recipes-router.js
@@ -2,7 +2,6 @@ import { Router } from "express";
 import validateBody from "../middleware/validateBody.js";
 import authenticate from "../middleware/authenticate.js";
 import recipesControllers from "../controllers/recipesControllers.js";
-import favoriteRecipesControllers from "../controllers/favoriteRecipesControllers.js";
 import { createRecipeSchema } from "../schemas/recipesSchemas.js";
 
 const validateRequestBody = validateBody(createRecipeSchema);
@@ -14,7 +13,7 @@ recipesRouter.get("/popular", recipesControllers.getPopularRecipes);
 
 recipesRouter.get("/favorite", authenticate, recipesControllers.getFavoriteRecipes);
 
-recipesRouter.delete(":id/favorite", authenticate, favoriteRecipesControllers.deleteFavoriteRecipe);
+recipesRouter.delete(":id/favorite", authenticate, recipesControllers.deleteFavoriteRecipe);
 
 recipesRouter.delete("/:id", authenticate, recipesControllers.deleteRecipe);
